Seguridad + borrar
Se ha introducido seguridad para evitar que x roles utilicen y comandos, y se ha añadido el comando borrar, para limpiar mensajes

+++ comandos/decir.js
@@ -1,5 +1,13 @@
 exports.run = (client, message, args) => {
-    let text = args.slice(0).join(" ");
-	message.delete();
-	message.channel.send(text, {tts: true}).catch(console.error);
+	if (args[0] != null) {
+		if(message.member.roles.some(r=>["Señor total del universo", "Alto rango del infierno", "Alta ranga dela infierna"].includes(r.name)) ) {
+			let text = args.slice(0).join(" ");
+			message.delete();
+			message.channel.send(text, {tts: true}).catch(console.error);
+		} else {
+			message.channel.send('No tienes privilegios suficientes').catch(console.error);
+		}
+	} else {
+		message.channel.send('No has introducido ningun mensaje').catch(console.error);
+	}
 }
